+++ server/lambda.ts
@@ -5,8 +5,9 @@
 export { handler as healthCheck } from './handlers/healthCheck';
 export { handler as getSpaces } from './handlers/getSpaces';
 export { handler as getSpaceById } from './handlers/getSpaceById';
-export { handler as getServices } from './handlers/getServices';
-export { handler as getSpaceServices } from './handlers/getSpaceServices';
-export { handler as getSpacePricing } from './handlers/getSpacePricing';
 export { handler as createReport } from './handlers/createReport';
 export { handler as handleOptions } from './handlers/handleOptions';
+
+// Blog handlers
+export { handler as getPosts } from './handlers/getPosts';
+export { handler as getPostBySlug } from './handlers/getPostBySlug';